pyproject.toml
[project]
name = "iot-ai-security-camera"
version = "0.1.0"
description = "Smart Security Camera for Verified User Access (IoT + edge face recognition)."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Christian Farrell" },
    { name = "Jeremy Frank" },
]
dependencies = [
    "paho-mqtt>=2.1.0,<3",
    "Flask>=3.0.3,<4",
    "python-dotenv>=1.0.1,<2",
    "pyserial>=3.5",
]

[project.optional-dependencies]
# Installed only on the Raspberry Pi gateway. Kept optional so a dev laptop
# can still `uv sync` without building dlib / picamera2.
pi = [
    "face-recognition>=1.3.0",
    "opencv-python-headless>=4.9",
]

[project.scripts]
# Full edge service: HTTP control plane + MQTT publisher + heartbeat in one process.
camera-service = "src.api_service:main"
# MQTT-only smoke test (heartbeat + manual publishes). Useful to verify the
# broker without bringing Flask up.
camera-mqtt = "src.mqtt_service:main"

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["artifacts", "captures", "docs"]

[tool.ruff.lint]
# Enable common Python linting rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "D",   # pydocstyle (docstring conventions)
]
ignore = [
    "E303",  # too many blank lines (we want 2)
    "E304",  # blank lines found after function decorator
    "W293",  # blank line contains whitespace (allow whitespace in blank lines, especially in comments/docstrings)
    "D100",  # Missing docstring in public module (optional)
    "D104",  # Missing docstring in public package (optional)
    "D105",  # Missing docstring in magic method (optional)
    "D107",  # Missing docstring in __init__ (optional)
    "D203",  # Incompatible with D211 (blank line before class)
    "D211",  # Incompatible with D203 (no blank line before class)
    "D212",  # Incompatible with D213 (multi-line summary first line)
    "D213",  # Incompatible with D212 (multi-line summary second line)
]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings
convention = "google"

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["src", "tests", "persistence", "picam", "services"]
combine-as-imports = true

[tool.ruff.format]
# Ruff formatter configuration
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
# Pytest configuration
pythonpath = ["."]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
